requirements.txt
absl-py==2.0.0
Brotli==1.0.9
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==2.0.4
contourpy==1.1.1
cryptography==41.0.7
cycler==0.12.1
einops==0.7.0
filelock==3.13.1
fonttools==4.47.0
gitdb==4.0.11
GitPython==3.1.41
gmpy2==2.1.2
google-auth==2.26.1
google-auth-oauthlib==1.0.0
grpcio==1.60.0
idna==3.4
imageio==2.33.1
imgaug==0.4.0
importlib-metadata==7.0.1
importlib-resources==6.1.1
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.7.4
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.9.0.80
packaging==23.2
pandas==2.0.3
pillow==10.2.0
pip==23.3.1
protobuf==4.25.1
psutil==5.9.8
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.2.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
setuptools==68.2.2
shapely==2.0.2
six==1.16.0
smmap==5.0.1
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.0
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2023.7.10
tqdm==4.66.1
triton==2.0.0
typing_extensions==4.7.1
tzdata==2023.4
ultralytics==8.1.11
urllib3==1.26.18
Werkzeug==3.0.1
wheel==0.41.2
zipp==3.17.0
